third_party/blink/web_tests/external/wpt/offscreen-canvas/text/2d.text.draw.align.center-expected.txt

This is a testharness.js-based test.
FAIL textAlign center is the center of the em squares (not the bounding box) assert_approx_equals: Red channel of the pixel at (5, 5) expected 0 +/- 2 but got 255
Harness: the test ran to completion.
